+++ src/ext_widgets/text_selector/text_selector.c
@@ -33,7 +33,8 @@
 
 static const wchar_t* text_selector_get_wtext(widget_t* widget);
 static ret_t text_selector_set_all_options_localize_text(widget_t* widget);
-static ret_t text_selector_sync_yoffset_with_selected_index(text_selector_t* text_selector);
+static ret_t text_selector_scroll_to(widget_t* widget, int32_t yoffset_end);
+static ret_t text_selector_sync_yoffset_with_selected_index(text_selector_t* text_selector, bool_t is_anim);
 
 const char* s_text_selector_properties[] = {WIDGET_PROP_TEXT,
                                             WIDGET_PROP_VALUE,
@@ -43,6 +44,7 @@ const char* s_text_selector_properties[] = {WIDGET_PROP_TEXT,
                                             WIDGET_PROP_LOCALIZE_OPTIONS,
                                             TEXT_SELECTOR_PROP_Y_SPEED_SCALE,
                                             TEXT_SELECTOR_PROP_LOOP_OPTIONS,
+                                            TEXT_SELECTOR_PROP_ANIMATION_TIME,
                                             NULL};
 
 static ret_t text_selector_paint_mask(widget_t* widget, canvas_t* c) {
@@ -265,7 +267,7 @@ static ret_t text_selector_on_layout_children(widget_t* widget) {
   return_value_if_fail(text_selector != NULL, RET_BAD_PARAMS);
   text_selector_set_real_draw_info(widget);
   if (!text_selector->is_init) {
-    text_selector_sync_yoffset_with_selected_index(text_selector);
+    text_selector_sync_yoffset_with_selected_index(text_selector, FALSE);
     text_selector->is_init = TRUE;
   }
   return RET_OK;
@@ -401,6 +403,9 @@ static ret_t text_selector_get_prop(widget_t* widget, const char* name, value_t*
   } else if (tk_str_eq(name, TEXT_SELECTOR_PROP_LOOP_OPTIONS)) {
     value_set_bool(v, text_selector->loop_options);
     return RET_OK;
+  } else if (tk_str_eq(name, TEXT_SELECTOR_PROP_ANIMATION_TIME)) {
+    value_set_uint32(v, text_selector->animating_time);
+    return RET_OK;
   }
 
   return RET_NOT_FOUND;
@@ -436,6 +441,8 @@ static ret_t text_selector_set_prop(widget_t* widget, const char* name, const va
     return text_selector_set_yspeed_scale(widget, value_float(v));
   } else if (tk_str_eq(name, TEXT_SELECTOR_PROP_LOOP_OPTIONS)) {
     return text_selector_set_loop_options(widget, value_bool(v));
+  } else if (tk_str_eq(name, TEXT_SELECTOR_PROP_ANIMATION_TIME)) {
+    return text_selector_set_animating_time(widget, value_uint32(v));
   }
 
   return RET_NOT_FOUND;
@@ -476,6 +483,7 @@ static ret_t text_selector_set_selected_index_only(text_selector_t* text_selecto
     value_set_uint32(&(evt.new_value), index);
 
     if (widget_dispatch(widget, (event_t*)&evt) != RET_STOP) {
+      text_selector->last_selected_index = text_selector->selected_index;
       text_selector->selected_index = index;
       if (widget->tr_text != NULL) {
         widget->tr_text = tk_str_copy(widget->tr_text, option->tr_text);
@@ -502,16 +510,47 @@ static ret_t text_selector_sync_selected_index_with_yoffset(text_selector_t* tex
   return text_selector_set_selected_index_only(text_selector, selected_index);
 }
 
-static ret_t text_selector_sync_yoffset_with_selected_index(text_selector_t* text_selector) {
+static ret_t text_selector_get_yoffset_for_selected_index(text_selector_t* text_selector, int32_t mid_index, int32_t item_height) {
+  int32_t yoffset = 0;
+  if (text_selector->loop_options) {
+    int32_t options_nr = text_selector_count_options(WIDGET(text_selector));
+    int32_t n = options_nr >> 1;
+    int32_t d = n - tk_abs(tk_abs(text_selector->selected_index - text_selector->last_selected_index) - n);
+    yoffset = text_selector->yoffset;
+    if (text_selector->selected_index > text_selector->last_selected_index) {
+      if (text_selector->selected_index - d == text_selector->last_selected_index) {
+        yoffset += (d * item_height);
+      } else {
+        yoffset -= (d * item_height);
+      }
+    } else {
+      if ((text_selector->last_selected_index + d + 1) % options_nr == text_selector->selected_index) {
+        yoffset += ((d + 1) * item_height);
+      } else {
+        yoffset -= (d * item_height);
+      }
+    }
+  } else {
+    yoffset = (text_selector->selected_index - mid_index) * item_height;
+  }
+  return yoffset;
+}
+
+static ret_t text_selector_sync_yoffset_with_selected_index(text_selector_t* text_selector, bool_t is_anim) {
   widget_t* widget = WIDGET(text_selector);
   int32_t mid_index = text_selector->visible_nr / 2;
   int32_t item_height = text_selector->draw_widget_h / text_selector->visible_nr;
-
-  text_selector->yoffset = (text_selector->selected_index - mid_index) * item_height;
-
-  widget_invalidate(widget, NULL);
-
-  return RET_OK;
+  if (is_anim && !widget->loading) {
+    int32_t yoffset = text_selector_get_yoffset_for_selected_index(text_selector, mid_index, item_height);
+    if (text_selector->wa != NULL) {
+      widget_animator_destroy(text_selector->wa);
+      text_selector->wa = NULL;
+    }
+    return text_selector_scroll_to(widget, yoffset);
+  } else {
+    text_selector->yoffset = (text_selector->selected_index - mid_index) * item_height;
+    return widget_invalidate(widget, NULL);
+  }
 }
 
 static ret_t text_selector_on_scroll_done(void* ctx, event_t* e) {
@@ -575,7 +614,7 @@ static ret_t text_selector_scroll_to(widget_t* widget, int32_t yoffset_end) {
   }
 
 #ifndef WITHOUT_WIDGET_ANIMATORS
-  text_selector->wa = widget_animator_scroll_create(widget, TK_ANIMATING_TIME, 0, EASING_SIN_INOUT);
+  text_selector->wa = widget_animator_scroll_create(widget, text_selector->animating_time, 0, EASING_SIN_INOUT);
   return_value_if_fail(text_selector->wa != NULL, RET_OOM);
 
   widget_animator_scroll_set_params(text_selector->wa, 0, yoffset, 0, yoffset_end);
@@ -621,7 +660,7 @@ static ret_t text_selector_up(widget_t* widget) {
   int32_t yoffset = 0;
   text_selector_t* text_selector = TEXT_SELECTOR(widget);
   return_value_if_fail(text_selector != NULL, RET_BAD_PARAMS);
-  yoffset = text_selector->yoffset - (widget->h / text_selector->visible_nr);
+  yoffset = text_selector->yoffset - (text_selector->draw_widget_h / text_selector->visible_nr);
 
   return text_selector_scroll_to(widget, yoffset);
 }
@@ -630,7 +669,7 @@ static ret_t text_selector_down(widget_t* widget) {
   int32_t yoffset = 0;
   text_selector_t* text_selector = TEXT_SELECTOR(widget);
   return_value_if_fail(text_selector != NULL, RET_BAD_PARAMS);
-  yoffset = text_selector->yoffset + (widget->h / text_selector->visible_nr);
+  yoffset = text_selector->yoffset + (text_selector->draw_widget_h / text_selector->visible_nr);
 
   return text_selector_scroll_to(widget, yoffset);
 }
@@ -689,7 +728,7 @@ static ret_t text_selector_on_event(widget_t* widget, event_t* e) {
     }
     case EVT_RESIZE:
     case EVT_MOVE_RESIZE: {
-      text_selector_sync_yoffset_with_selected_index(text_selector);
+      text_selector_sync_yoffset_with_selected_index(text_selector, FALSE);
       break;
     }
     default:
@@ -735,6 +774,7 @@ widget_t* text_selector_create(widget_t* parent, xy_t x, xy_t y, wh_t w, wh_t h)
   text_selector->yspeed_scale = 1.0f;
 
   text_selector->locale_info_id = TK_INVALID_ID;
+  text_selector->animating_time = TK_ANIMATING_TIME;
 
   return widget;
 }
@@ -908,8 +948,10 @@ ret_t text_selector_set_selected_index(widget_t* widget, uint32_t index) {
   return_value_if_fail(text_selector != NULL, RET_BAD_PARAMS);
   return_value_if_fail(option != NULL, RET_BAD_PARAMS);
 
-  text_selector_set_selected_index_only(text_selector, index);
-  text_selector_sync_yoffset_with_selected_index(text_selector);
+  if (index != text_selector->selected_index) {
+    text_selector_set_selected_index_only(text_selector, index);
+    text_selector_sync_yoffset_with_selected_index(text_selector, TRUE);
+  }
 
   return widget_invalidate(widget, NULL);
 }
@@ -923,7 +965,7 @@ ret_t text_selector_set_visible_nr(widget_t* widget, uint32_t visible_nr) {
   } else {
     text_selector->visible_nr = 1;
   }
-  text_selector_sync_yoffset_with_selected_index(text_selector);
+  text_selector_sync_yoffset_with_selected_index(text_selector, FALSE);
 
   return widget_invalidate(widget, NULL);
 }
@@ -1005,6 +1047,15 @@ ret_t text_selector_set_yspeed_scale(widget_t* widget, float_t yspeed_scale) {
   return_value_if_fail(text_selector != NULL, RET_BAD_PARAMS);
 
   text_selector->yspeed_scale = yspeed_scale;
+
+  return RET_OK;
+}
+
+ret_t text_selector_set_animating_time(widget_t* widget, uint32_t animating_time) {
+  text_selector_t* text_selector = TEXT_SELECTOR(widget);
+  return_value_if_fail(text_selector != NULL, RET_BAD_PARAMS);
+
+  text_selector->animating_time = animating_time;
 
   return RET_OK;
 }